requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
langchain_community
streamlit






